moved Imanee\FilterInterface to Imanee\Model\FilterInterface
