+++ src/Filter/BWFilter.php
@@ -2,7 +2,7 @@
 
 namespace Imanee\Filter;
 
-use Imanee\FilterInterface;
+use Imanee\Model\FilterInterface;
 
 /**
  * Black and White Filter
